Move common parser to the main script and delete module
