+++ download.py
@@ -6,11 +6,19 @@ import urllib2
 import zipfile
 import zlib
 
+from argparse import ArgumentParser
 from cStringIO import StringIO
 
-from common import parse_args
-
 cwd = os.path.dirname(os.path.realpath(__file__))
+
+
+def parse_args():
+    parser = ArgumentParser()
+    parser.add_argument("--url", dest="url",
+                        help='File to download and unpack.')
+    parser.add_argument("--times", dest="times", default=1, type=int,
+                        help='How many times to test download a file.')
+    return parser.parse_args()
 
 
 # Not being used at the moment
